refactor: removed save_local from wikipedia_articles_embedding_store
