+++ etl/models/__init__.py
@@ -1,3 +1,3 @@
+from .record import Record as Record  # isort:skip
 from .document_tuple import DocumentTuple as DocumentTuple
-from .record import Record as Record
 from .record_tuple import RecordTuple as RecordTuple
