+++ etl/pipelines/__init__.py
@@ -2,9 +2,10 @@ from .embedding_pipeline import EmbeddingPipeline as EmbeddingPipeline
 from .open_ai_embedding_pipeline import (
     OpenAiEmbeddingPipeline as OpenAiEmbeddingPipeline,
 )
-from .open_ai_record_enrichment_pipeline import (
-    OpenAiRecordEnrichmentPipeline as OpenAiRecordEnrichmentPipeline,
-)
 from .record_enrichment_pipeline import (
     RecordEnrichmentPipeline as RecordEnrichmentPipeline,
+)
+
+from .open_ai_record_enrichment_pipeline import (  # isort:skip
+    OpenAiRecordEnrichmentPipeline as OpenAiRecordEnrichmentPipeline,
 )
