+++ etl/pipelines/embedding_pipeline.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from typing import final
 
 from langchain.docstore.document import Document
-from langchain_community.vectorstores import FAISS
+from langchain_community.vectorstores import FAISS, VectorStore
 from langchain_core.embeddings import Embeddings
 
 
@@ -14,7 +14,7 @@ class EmbeddingPipeline(ABC):
         """Return an embedding model that will be used to create an embedding store."""
 
     @final
-    def create_embedding_store(self, documents: tuple[Document, ...]) -> FAISS:
+    def create_embedding_store(self, documents: tuple[Document, ...]) -> VectorStore:
         """Return an embedding store that contains embeddings of Documents."""
 
         return FAISS.from_documents(
