+++ etl/pipelines/open_ai_record_enrichment_pipeline.py
@@ -7,7 +7,7 @@ from langchain_openai import ChatOpenAI
 
 from etl.models import Record, wikipedia
 from etl.models.types import EnrichmentType, ModelQuestion, ModelResponse, RecordKey
-from etl.pipelines.record_enrichment_pipeline import RecordEnrichmentPipeline
+from etl.pipelines import RecordEnrichmentPipeline
 from etl.resources import OpenAiPipelineConfig
