+++ etl/resources/__init__.py
@@ -1,4 +1,7 @@
 from .input_data_files_config import InputDataFilesConfig as InputDataFilesConfig
-from .open_ai_pipeline_config import OpenAiPipelineConfig as OpenAiPipelineConfig
 from .open_ai_settings import OpenAiSettings as OpenAiSettings
 from .output_config import OutputConfig as OutputConfig
+
+from .open_ai_pipeline_config import (  # isort:skip
+    OpenAiPipelineConfig as OpenAiPipelineConfig,
+)
